requirements.txt
python-docx
transformers
sentence-transformers
faiss-cpu
huggingface-hub
numpy
torch
requests
streamlit